+++ atui/SearchService/src/main.js
@@ -124,3 +124,9 @@ function atuiSearchservices_HeaderPropositionsGenerate() {
         }
     } catch {}
 }
+
+/* Prevent default <form> submission */
+
+document.querySelector("#atuiSearchservices_HeaderBar form").addEventListener("submit", function (event) {
+    event.preventDefault();
+});